move backend interface to src/backend
